+++ flutter/plugins/preferences_local_storage_inspector/example/example.dart
@@ -1,9 +1,13 @@
 import 'package:flutter/material.dart';
+import 'package:preferences_local_storage_inspector/preferences_local_storage_inspector.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import 'package:storage_inspector/storage_inspector.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
+
+  // ignore: avoid_print
+  storageInspectorLogger = (e) => print(e);
 
   final preferences = await SharedPreferences.getInstance();
 
@@ -19,18 +23,11 @@ void main() async {
   });
   driver.addKeyValueServer(keyValueServer);
 
-  final fileServer = DefaultFileServer(_documentsDirectory(), 'App Documents');
-  driver.addFileServer(fileServer);
-
   // Don't wait for a connection from the instrumentation driver
   await driver.start(paused: false);
 
   // run app
+  await Future<void>.delayed(const Duration(minutes: 15));
 
   await driver.stop(); //Optional when main ends
-}
-
-//Use path_provider to provide this
-String _documentsDirectory() {
-  return '.';
 }