+++ flutter/plugins/secure_storage_local_storage_inspector/lib/src/secure_storage_key_value_server.dart
@@ -1,5 +1,4 @@
 import 'package:flutter_secure_storage/flutter_secure_storage.dart';
-import 'package:storage_inspector/src/servers/simple/simple_key_value_server.dart';
 import 'package:storage_inspector/storage_inspector.dart';
 import 'package:uuid/uuid.dart';